V2.3 of updating Model to match DA standards
